backend/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Caching & Sessions
redis==5.0.1
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# ML & Data Science
lightgbm==4.1.0
scikit-learn==1.3.2
pandas==2.2.0
numpy==1.26.4
joblib==1.3.2

# Network Security
scapy==2.5.0
python-nmap==0.7.1
yara-python==4.5.4

# Background Tasks
celery==5.3.4
flower==2.0.1

# WebSocket & Real-time
python-socketio==5.10.0
websockets==12.0

# HTTP & API
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Security
cryptography==42.0.5
python-dotenv==1.0.0
bcrypt==4.1.2

# Blockchain
# hashlib2==1.0.1  # Removed - not available for Python 3.12

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
